requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
anyio==3.7.1
apache-airflow==2.7.0
apache-airflow-providers-common-sql==1.14.2
apache-airflow-providers-ftp==3.5.0
apache-airflow-providers-google==10.21.0
apache-airflow-providers-http==4.5.0
apache-airflow-providers-imap==3.3.0
apache-airflow-providers-sqlite==3.4.3
apispec==6.3.0
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backoff==1.10.0
blinker==1.6.2
cachelib==0.9.0
cachetools==5.4.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron_descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.7
db-dtypes==1.2.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.4.2
docstring_parser==0.16
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.3
Flask==2.2.5
Flask-AppBuilder==4.3.3
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.4.0
fsspec==2024.6.1
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.2.0
gcsfs==2024.6.1
google-ads==24.1.0
google-analytics-admin==0.22.9
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.60.0
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.2.5
google-cloud-automl==2.13.5
google-cloud-batch==0.17.23
google-cloud-bigquery==3.20.1
google-cloud-bigquery-datatransfer==3.15.5
google-cloud-bigtable==2.25.0
google-cloud-build==3.24.2
google-cloud-compute==1.19.2
google-cloud-container==2.50.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.20.1
google-cloud-dataflow-client==0.8.12
google-cloud-dataform==0.5.11
google-cloud-dataplex==2.2.2
google-cloud-dataproc==5.10.2
google-cloud-dataproc-metastore==1.15.5
google-cloud-dlp==3.20.0
google-cloud-kms==2.24.2
google-cloud-language==2.14.0
google-cloud-logging==3.11.0
google-cloud-memcache==1.9.5
google-cloud-monitoring==2.22.2
google-cloud-orchestration-airflow==1.13.1
google-cloud-os-login==2.14.6
google-cloud-pubsub==2.23.0
google-cloud-redis==2.15.5
google-cloud-resource-manager==1.12.5
google-cloud-run==0.10.8
google-cloud-secret-manager==2.20.2
google-cloud-spanner==3.48.0
google-cloud-speech==2.27.0
google-cloud-storage==2.18.0
google-cloud-storage-transfer==1.11.5
google-cloud-tasks==2.16.5
google-cloud-texttospeech==2.16.5
google-cloud-translate==3.15.5
google-cloud-videointelligence==2.13.5
google-cloud-vision==3.7.4
google-cloud-workflows==1.14.5
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.7.1
googleapis-common-protos==1.60.0
graphviz==0.20.1
greenlet==2.0.2
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.65.4
grpcio-gcp==0.2.2
grpcio-status==1.62.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
json-merge-patch==0.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
looker-sdk==24.12.1
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.3.0
multidict==6.0.4
mysql-connector-python==9.0.0
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
ordered-set==4.1.0
packaging==23.1
pandas==2.1.4
pandas-gbq==0.23.1
pathspec==0.11.2
pendulum==2.1.2
pluggy==1.2.0
prison==0.2.1
proto-plus==1.24.0
protobuf==4.25.4
psutil==5.9.5
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydata-google-auth==1.8.2
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.5.2
rich-argparse==1.2.0
rpds-py==0.9.2
rsa==4.9
setproctitle==1.3.2
shapely==2.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
sqlalchemy-bigquery==1.11.0
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-spanner==1.7.0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2024.1
uc-micro-py==1.0.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.2
